Struct zip_codes::Record

pub struct Record {
    pub zip_code: &'static str,
    pub zip_code_type: Type,
    pub city: &'static str,
    pub state: &'static str,
    pub coordinates: Option<(f64, f64)>,
    pub is_decommissioned: bool,
    pub tax_returns_filed: Option<u64>,
    pub estimated_population: Option<u64>,
    pub total_wages: Option<u64>,
}

Information for a single ZIP code.

Fields

zip_code

The 5-digit ZIP code.

zip_code_type

The ZIP code classification.

city

The city to which the ZIP code belongs (all uppercase).

state

The state to which the ZIP code belongs (two letter abbreviation).

coordinates

Latitude and longitude

is_decommissioned

If true, the ZIP code is historical; if false, the ZIP code is current.

tax_returns_filed

The number of individual tax returns filed as of 2008.

estimated_population

The estimated population of the area of the ZIP code.

total_wages

Total yearly wages of the population.